Rename modules to match component names
